AGV (The Meatloaf) -Fixed the exec commands to run scripts properly -Added correct logic for determining whether data was received or sent

+++ mqtt_controller.c
@@ -16,7 +16,7 @@
 
 	 
 
-	 BUILD: gcc -o mqtt_controller mqtt_contller.c -lmosquitto
+	 BUILD: gcc -o mqtt_controller mqtt_controller.c -lmosquitto
 */
 
 #include <signal.h>
@@ -34,7 +34,8 @@
 // #define mqtt_port 1883
 
 // roborio
-#define mqtt_host "10.23.58.2"
+// #define mqtt_host "10.23.58.2"
+#define mqtt_host "10.23.58.26"
 #define mqtt_port 1183
 
 static int run = 1;
@@ -61,6 +62,11 @@ const char *INVALID = "INVALID";
 const char *OFF = "off";
 const char *VIEW = "view";
 const char *VISION = "vision";
+
+// which camera is this?
+int claw_f = 0;
+int cargo_f = 0;
+
 
 enum processes cmd_check(char * command)
 {
@@ -120,7 +126,11 @@ void claw_start_cv()
 	case 0:
 		// child
 		// exec new process
-		execl("useless.sh", "useless.sh", "vision", (char *) NULL);
+		//execl("useless.sh", "useless.sh", "vision", (char *) NULL);
+		
+
+		execl("vision_startup.sh","vision_startup.sh", (char *) NULL);
+
 		// never reach this point
 		
 	default:
@@ -148,7 +158,52 @@ void claw_start_view()
 	case 0:
 		// child
 		// exec new process
-		execl("useless.sh", "useless.sh", "view", (char *) NULL);
+		// execl("useless.sh", "useless.sh", "view", (char *) NULL);
+
+
+		execl("/usr/bin/uv4l", "uv4l", "--auto-video_nr", "--driver", "raspicam",
+		 "--encoding", "h264", "--width", "104", "--height", "96",
+		 "--server-option", "--port=1187", "--framerate",
+		 "25", "--enable-server", "on", "--rotation", "180", (char *) NULL);
+
+
+		/*
+		// this works
+		execl("/usr/bin/uv4l", "uv4l", "--auto-video_nr", "--driver", "raspicam",
+		 "--encoding", "h264", 
+		 "--enable-server", "on", (char *) NULL);
+		*/
+
+
+		// never reach this point
+		
+	default:
+		// parent
+		claw_process.pid = ret;
+		break;
+	}
+}
+// this function starts computer vision by forking
+// the parent returns from the fork call with the pid of the child
+// the child returns with a zero and launches the task we wish to start
+void claw_kill_view()
+{
+	pid_t ret;
+	
+	ret = fork();
+	printf("fork returned %d\n", ret);
+	switch(ret) {
+	case -1:
+		// error
+		printf("unable to fork - %s\n", strerror(errno));
+		exit(-1);
+
+	case 0:
+		// child
+		// exec new process
+		execl("/usr/bin/pkill", "pkill", "uv4l", (char *) NULL);
+
+
 		// never reach this point
 		
 	default:
@@ -179,8 +234,12 @@ void claw_process_change(char *command)
 	if(claw_process.current_process != off) {
 		printf("Kill claw pid %d\n", claw_process.pid);
 		if(claw_process.pid != 0) {
-			ret = kill(claw_process.pid, SIGTERM);
-			printf("Kill returned %d\n", ret);
+			if(claw_process.current_process == view) {
+				claw_kill_view();
+			} else {
+				ret = kill(claw_process.pid, SIGTERM);
+				printf("Kill returned %d\n", ret);
+			}
 		}
 		claw_process.current_process = off;
 	}
@@ -200,10 +259,115 @@ void claw_process_change(char *command)
 	
 }
 
+
+// this function starts computer vision by forking
+// the parent returns from the fork call with the pid of the child
+// the child returns with a zero and launches the task we wish to start
+void cargo_start_cv()
+{
+	pid_t ret;
+	
+	ret = fork();
+	printf("fork returned %d\n", ret);
+	switch(ret) {
+	case -1:
+		// error
+		printf("unable to fork - %s\n", strerror(errno));
+		exit(-1);
+
+	case 0:
+		// child
+		// exec new process
+		execl("useless.sh", "useless.sh", "vision", (char *) NULL);
+		// never reach this point
+		
+	default:
+		// parent
+		cargo_process.pid = ret;
+		break;
+	}
+}
+
+
+// this function starts remote viewing by forking
+// the parent returns from the fork call with the pid of the child
+// the child returns with a zero and launches the task we wish to start
+void cargo_start_view()
+{
+	pid_t ret;
+	
+	ret = fork();
+	printf("fork returned %d\n", ret);
+	switch(ret) {
+	case -1:
+		// error
+		printf("unable to fork - %s\n", strerror(errno));
+		exit(-1);
+
+	case 0:
+		// child
+		// exec new process
+		// execl("useless.sh", "useless.sh", "view", (char *) NULL);
+
+
+		execl("/usr/bin/uv4l", "uv4l", "--auto-video_nr", "--driver", "raspicam",
+		 "--encoding", "h264", "--width", "104", "--height", "96",
+		 "--server-option", "--port=1187", "--framerate",
+		 "25", "--enable-server", "on", "--rotation", "180", (char *) NULL);
+
+
+		/*
+		// this works
+		execl("/usr/bin/uv4l", "uv4l", "--auto-video_nr", "--driver", "raspicam",
+		 "--encoding", "h264", 
+		 "--enable-server", "on", (char *) NULL);
+		*/
+
+
+		// never reach this point
+		
+	default:
+		// parent
+		cargo_process.pid = ret;
+		break;
+	}
+}
+// this function kills remote viewing by forking
+// the parent returns from the fork call with the pid of the child
+// the child returns with a zero and launches the task we wish to start
+void cargo_kill_view()
+{
+	pid_t ret;
+	
+	ret = fork();
+	printf("fork returned %d\n", ret);
+	switch(ret) {
+	case -1:
+		// error
+		printf("unable to fork - %s\n", strerror(errno));
+		exit(-1);
+
+	case 0:
+		// child
+		// exec new process
+		execl("/usr/bin/pkill", "pkill", "uv4l", (char *) NULL);
+
+
+		// never reach this point
+		
+	default:
+		// parent
+		cargo_process.pid = ret;
+		break;
+	}
+}
+
+
 void cargo_process_change(char *command)
 {
 	enum processes cmd;
-
+	int ret;
+	
 	cmd = cmd_check(command);
 	if(cmd == invalid) {
 		return;
@@ -212,13 +376,21 @@ void cargo_process_change(char *command)
 	// are we already running this command?
 	if(cargo_process.current_process == cmd) {
 		// nothing to do
-		printf("Cargo already in %s\n", str_cmd(cmd));
+		printf("cargo already in %s\n", str_cmd(cmd));
 		return;
 	}
 
 	// are we running something else that must be shutdown?
 	if(cargo_process.current_process != off) {
 		printf("Kill cargo pid %d\n", cargo_process.pid);
+		if(cargo_process.pid != 0) {
+			if(cargo_process.current_process == view) {
+				cargo_kill_view();
+			} else {
+				ret = kill(cargo_process.pid, SIGTERM);
+				printf("Kill returned %d\n", ret);
+			}
+		}
 		cargo_process.current_process = off;
 	}
 	if(cmd == off) {
@@ -227,11 +399,18 @@ void cargo_process_change(char *command)
 	}
 	
 	// start new process
+	if(cmd == vision) {
+		cargo_start_cv();
+	} else if(cmd == view) {
+		cargo_start_view();
+	}
 	cargo_process.current_process = cmd;
-	cargo_process.pid = (int) random();
 	printf("Starting new cargo process pid %d\n", cargo_process.pid);
 	
 }
+
+
+
 
 void handle_signal(int s)
 {
@@ -250,17 +429,25 @@ void message_callback(struct mosquitto *mosq, void *obj, const struct mosquitto_
 	printf("got message '%.*s' for topic '%s'\n", message->payloadlen, (char*) message->payload, message->topic);
 
 	mosquitto_topic_matches_sub("/camera/controls/claw/+", message->topic, &match);
-	if(match) {
+	if(claw_f && match) {
 		printf("got message for claw camera controls topic\n");
 		claw_process_change(message->payload);
 	}
 
 	mosquitto_topic_matches_sub("/camera/controls/cargo/+", message->topic, &match);
-	if(match) {
+	if(cargo_f && match) {
 		printf("got message for cargo camera controls topic\n");
 		cargo_process_change(message->payload);
 	}
 
+}
+
+void usage()
+{
+	printf("usage: mqtt_controller <camera choice>\n");
+	printf("where <camera choice> is either:\n");
+	printf("	claw\n");
+	printf("	cargo\n");
 }
 
 int main(int argc, char *argv[])
@@ -269,6 +456,23 @@ int main(int argc, char *argv[])
 	char clientid[24];
 	struct mosquitto *mosq;
 	int rc = 0;
+
+
+	if(argc != 2) {
+		usage();
+		exit(-1);
+	}
+
+	if(strcmp(argv[1], "claw") == 0) {
+		claw_f = 1;
+		cargo_f = 0;
+	} else if(strcmp(argv[1], "cargo") == 0) {
+		claw_f = 0;
+		cargo_f = 1;
+	} else {
+		usage();
+		exit(-1);
+	}
 
 	signal(SIGINT, handle_signal);
 	signal(SIGTERM, handle_signal);
